Made email model
